+++ src/NBitcoin.Tests/transaction_tests.cs
@@ -3346,6 +3346,25 @@ namespace NBitcoin.Tests
             Assert.True(!StandardScripts.IsStandardTransaction(t));
         }
 
+        [Fact]
+        public void CanDecodeAndEncodeRawTransaction()
+        {
+            var tests = TestCase.read_json("data/tx_raw.json");
+            foreach (var test in tests)
+            {
+                var format = (RawFormat)Enum.Parse(typeof(RawFormat), (string)test[0], true);
+                var network = ((string)test[1]) == "Main" ? Network.Main : Network.TestNet;
+                var testData = ((JObject)test[2]).ToString();
+
+                Transaction raw = Transaction.Parse(testData, format, network);
+
+                TestUtils.AssertJsonEquals(raw.ToString(format, network), testData);
+
+                var raw3 = Transaction.Parse(raw.ToString(format, network), format);
+                Assert.Equal(raw.ToString(format, network), raw3.ToString(format, network));
+            }
+        }
+
         private byte[] ParseHex(string data)
         {
             return Encoders.Hex.DecodeData(data);
