+++ src/Stratis.Bitcoin.IntegrationTests/Data/invalid.cookie
@@ -1,0 +1,1 @@
+__not_a_valid_cookie__:3433
